Split Rpc code
